requirements.txt
Flask
flask-cors
python-dotenv
requests
beautifulsoup4
google-api-python-client
google-generativeai